Store the wall length / 2
Since the only place that it is used it is divided by two.
This also means that for the longwall.wad the length easily fits in an unsigned 32 bit number (the longest length is now 0x40000000), so reduce the size of the variable to suit.

+++ Source/r_defs.h
@@ -322,7 +322,7 @@ typedef struct
     // backsector is NULL for one sided lines
     sector_t*	frontsector;
     sector_t*	backsector;
-    int64_t	length;
+    uint32_t	length;
 } seg_t;
